requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-extensions==3.2.3
psycopg2-binary==2.9.7
python-decouple==3.8
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-ratelimit==4.1.0
django-silk==5.0.4
django-debug-toolbar==4.2.0
django-storages==1.14.2
boto3==1.34.70
botocore==1.34.70
aiobotocore==2.13.3
whitenoise==6.6.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2
black==23.9.1
flake8==6.1.0
pylint==2.17.5
isort==5.12.0
pre-commit==3.5.0
django-environ==0.11.2
django-cleanup==9.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.2.0
django-allauth==0.57.0
django-oauth-toolkit==1.7.1
drf-spectacular==0.26.5
django-health-check==3.17.0
sentry-sdk==1.38.0
django-prometheus==2.3.1
